requirements.txt
flask
replicate
uunicorn
google-genai
openai
markdown
